refactor usebeforeunload into a separate hook
