+++ frontend/javascripts/dashboard/dataset/useBeforeUnload_hook.ts
@@ -1,0 +1,54 @@
+import type {
+  Action as HistoryAction,
+  Location as HistoryLocation,
+  UnregisterCallback,
+} from "history";
+import { useCallback, useEffect, useRef } from "react";
+import { useHistory } from "react-router-dom";
+
+const useBeforeUnload = (hasUnsavedChanges: boolean, message: string) => {
+  const history = useHistory();
+  const unblockRef = useRef<UnregisterCallback | null>(null);
+  const blockTimeoutIdRef = useRef<number | null>(null);
+
+  const unblockHistory = useCallback(() => {
+    window.onbeforeunload = null;
+    if (blockTimeoutIdRef.current != null) {
+      clearTimeout(blockTimeoutIdRef.current);
+      blockTimeoutIdRef.current = null;
+    }
+    if (unblockRef.current != null) {
+      unblockRef.current();
+      unblockRef.current = null;
+    }
+  }, []);
+
+  useEffect(() => {
+    const beforeUnload = (
+      newLocation: HistoryLocation<unknown>,
+      action: HistoryAction,
+    ): string | false | void => {
+      if (action === undefined || newLocation.pathname !== window.location.pathname) {
+        if (hasUnsavedChanges) {
+          window.onbeforeunload = null;
+          blockTimeoutIdRef.current = window.setTimeout(() => {
+            // @ts-ignore
+            window.onbeforeunload = beforeUnload;
+          }, 500);
+          return message;
+        }
+      }
+      return;
+    };
+
+    unblockRef.current = history.block(beforeUnload);
+    // @ts-ignore
+    window.onbeforeunload = beforeUnload;
+
+    return () => {
+      unblockHistory();
+    };
+  }, [history, hasUnsavedChanges, message, unblockHistory]);
+};
+
+export default useBeforeUnload;
